fix: share의 webToken 저장 관련 로직이 빠져있던 문제 해결
