+++ android/app/build.gradle
@@ -137,8 +137,8 @@ android {
         applicationId "kr.ygtang"
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
-        versionCode 4
-        versionName "1.0.1-android.1"
+        versionCode 5
+        versionName "1.0.1-android.2"
         buildConfigField "boolean", "IS_NEW_ARCHITECTURE_ENABLED", isNewArchitectureEnabled().toString()
 
         if (isNewArchitectureEnabled()) {
